Customize mouse input to conform with EFS1.4
To deal with #48 we try this fix where a certain amount of space/time
uncertainty is allowed with respect to registering mouse clicks on the
Planet/Space maps. This breaks 20 out of 34 automated tests which is 
expected. We do not fix the tests yet, as this change needs some general
testing before it can be accepted.



src/gui/Gui.java