Boot time errors now report to boot GUI
There seems to be some trouble with Throwables in the early boot stage
so boot time failures will be reported as text only to the boot GUI and
System.out. Input logging is also started after the main Gui is ready
to be set visible and the boot GUI is disposed of so that the boot GUI
does not generate Events/Throwables.



src/gui/Gui.java
src/phoenix/Phoenix.java
src/util/CrashReporter.java
src/util/FileNameCapitalizer.java